Extracted flatlist from AllBookings to be a separate component - BookingsList
Created a BookingsCalendar component and added to AllBookings. Shows bookings in Calendar format
User can toggle between list view or calendar view by pressing button - to be implemented
